Major clean-up of old features
This is breaking change. The most glaring removal is deepdish.plot,
which has now been migrated to vzlog.image. This removal will allow
deepdish to be easily distributed (without requiring cython).

The rest were hidden features that did not have proper documentation, so
it should not break for any users. Some of these features may be useful
(such as concat and whiten) and might be re-introduced at some point.