Added support for pandas data structures
DataFrame, Series and Panel are all supported.

Since pandas uses PyTables as backend, it has been pretty painless to
add support for it. All it took for saving and loading was a subclass
of HDFStore that can attach an already established handler. The
inspection of the shape relies on the internals and could break as
pandas is updated. However, but this code has been made so that it can
silently break without much fuss.