MAJOR: Removed the unpack option from io.load
This feature was a bit confusing, since depending on it and then adding
another entry to a dictionary is strange behavior. Instead, deepdish now
marks fake top-level /data groups so that it knows whether or not to
unpack them. It doesn't hurt to be explicit when loading hdf5 files
simpy are "named" entries:

    dd.io.load('feat.h5', '/features')

This commit is marked major, since it could break programs that use or
rely on the previous behavior.